chapter-4 : Add Bin filling management with asynchronous communications using Kafka

+++ business/bin-service/src/main/java/com/binwatcher/binservice/controller/BinController.java
@@ -47,7 +47,7 @@ public class BinController {
     }
 
     @PatchMapping("/{id}/fill-level")
-    public ResponseEntity<Bin> updateFillLevel(@PathVariable String id, @RequestParam Integer level) {
+    public ResponseEntity<Bin> updateFillLevel(@PathVariable String id, @RequestParam short level) {
         try {
             Bin bin = binService.updateFillLevel(id, level);
             return new ResponseEntity<>(bin, HttpStatus.OK);
